feat: CI/CD and update documentation

+++ example_test.go
@@ -18,12 +18,13 @@ func ExampleNew() {
 	}
 
 	w, h := 100, 100
-	heightmap := make([]float64, w*h)
+
+	heightmap := make([]float64, int(w*h))
 
 	for y := 0; y < h; y++ {
 		for x := 0; x < w; x++ {
-			xFloat := float64(x) / float64(w)
-			yFloat := float64(y) / float64(h)
+			xFloat := float64(x / w)
+			yFloat := float64(y / h)
 
 			heightmap[(y*w)+x] = noise.Eval64(xFloat, yFloat)
 		}
